PB/club/requirements.txt
django
djangorestframework
django-recurrence
djangorestframework-simplejwt
Pillow
python-dateutil
django-cors-headers